+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- [BREAKING CHANGE] The sandbox manager's `withSandbox` method now requires an array of nodes from the original tree and the corresponding sandbox nodes are now positional arguments of the callback function.
+
 ## 0.48.5
 
 - Fixed incompatibility with MobX >= 6.1 (in order to fix it `$modelId` is now an enumerable property though).